Fix incorrect param in Canvas.create_poll. Relocate test to appropriate file. Update Changelog (#471)
